requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.0.2
pillow==11.0.0
proglog==0.1.10
requests==2.32.3
tqdm==4.66.6
urllib3==2.2.3
